Fix linters errror
